insurance-policy-mapper/requirements/dev.txt
audioread==3.0.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
gdown==5.2.0
huggingface-hub==0.26.3
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
nltk==3.9.1
noisereduce==3.0.3
numba==0.60.0
numpy==2.0.2
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
pycparser==2.22
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
sympy==1.13.1
# Theano==1.0.5
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
torchaudio==2.5.1
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
urllib3==2.2.3
